requirements.txt
django[argon2]
djangorestframework
cryptography
djangorestframework-simplejwt[crypto]
psycopg2
Pillow